refactor(invoices): standardize invoice handling and enhance error management across services
